tags_to_df shifted to utils
