--- examples/extract_data.py
@@ -1,6 +1,0 @@
-# %%
-from miluv.data import DataLoader
-
-mv = DataLoader("1c", barometer=False)
-a=2
-# %%
