+++ miluv/utils.py
@@ -1,13 +1,7 @@
-import matplotlib.pyplot as plt
-import seaborn as sns
 import pandas as pd
 import yaml
 from os.path import join
-import numpy as np
-from csaps import csaps
-from scipy.spatial.transform import Rotation
-import scipy as sp
-from typing import Tuple
+from typing import Any
 
 def get_experiment_info(path):
     exp_name = path.split('/')[-1]
@@ -30,3 +24,60 @@ def get_tags(flatten=False):
         return arms
     else:
         return moment_arms
+
+def tags_to_df(anchors: Any = None, 
+               moment_arms: Any = None, 
+               april_tags: Any = None) -> pd.DataFrame:
+
+    uwb_tags = None
+    april_tags = None
+
+    if anchors is not None:
+        tags = []
+        for key, value in anchors.items():
+            # parse value
+            value = value.strip('[]').split(',')
+            tags.append({
+                'tag_id': int(key),
+                'parent_id': 'world',
+                'position.x': float(value[0]),
+                'position.y': float(value[1]),
+                'position.z': float(value[2]),
+            })
+        # Convert the data dictionary to a DataFrame
+        uwb_tags = pd.DataFrame(tags)
+    
+    if moment_arms is not None:
+        tags = []
+        for robot, arms in moment_arms.items():
+            for tag, value in arms.items():
+                value = value.strip('[]').split(',')
+                tags.append({
+                    'tag_id': int(tag),
+                    'parent_id': robot,
+                    'position.x': float(value[0]),
+                    'position.y': float(value[1]),
+                    'position.z': float(value[2]),
+                })
+        tags = pd.DataFrame(tags)
+    if uwb_tags is not None:
+        uwb_tags = pd.concat([uwb_tags, tags], 
+                                    ignore_index=True)
+    else:
+        uwb_tags = tags
+    
+    if april_tags is not None:
+        tags = []
+        for key, value in april_tags.items():
+            # parse value
+            value = value.strip('[]').split(',')
+            tags.append({
+                'tag_id': int(key),
+                'parent_id': 'world',
+                'position.x': float(value[0]),
+                'position.y': float(value[1]),
+                'position.z': float(value[2]),
+            })
+        # Convert the data dictionary to a DataFrame
+        april_tags = pd.DataFrame(tags)
+    return uwb_tags, april_tags
