+++ miluv/wrappers.py
@@ -9,65 +9,6 @@ from scipy.spatial.transform import Rotation
 This module contains data wrappers for:
 - MocapTrajectory
 """
-
-def tags_to_df(anchors: Any = None, 
-               moment_arms: Any = None, 
-               april_tags: Any = None) -> pd.DataFrame:
-
-    uwb_tags = None
-    april_tags = None
-
-    if anchors is not None:
-        tags = []
-        for key, value in anchors.items():
-            # parse value
-            value = value.strip('[]').split(',')
-            tags.append({
-                'tag_id': int(key),
-                'parent_id': 'world',
-                'position.x': float(value[0]),
-                'position.y': float(value[1]),
-                'position.z': float(value[2]),
-            })
-        # Convert the data dictionary to a DataFrame
-        uwb_tags = pd.DataFrame(tags)
-    
-    if moment_arms is not None:
-        tags = []
-        for robot, arms in moment_arms.items():
-            for tag, value in arms.items():
-                value = value.strip('[]').split(',')
-                tags.append({
-                    'tag_id': int(tag),
-                    'parent_id': robot,
-                    'position.x': float(value[0]),
-                    'position.y': float(value[1]),
-                    'position.z': float(value[2]),
-                })
-        tags = pd.DataFrame(tags)
-    if uwb_tags is not None:
-        uwb_tags = pd.concat([uwb_tags, tags], 
-                                    ignore_index=True)
-    else:
-        uwb_tags = tags
-    
-    if april_tags is not None:
-        tags = []
-        for key, value in april_tags.items():
-            # parse value
-            value = value.strip('[]').split(',')
-            tags.append({
-                'tag_id': int(key),
-                'parent_id': 'world',
-                'position.x': float(value[0]),
-                'position.y': float(value[1]),
-                'position.z': float(value[2]),
-            })
-        # Convert the data dictionary to a DataFrame
-        april_tags = pd.DataFrame(tags)
-    return uwb_tags, april_tags
-
-
 class MocapTrajectory:
     """
     This class provides a spline representation
